eda_agent/requirements.txt
# Core dependencies
langgraph>=1.0.1
langgraph-checkpoint-sqlite>=2.0.10
langchain-openai>=0.1.0
langchain-core>=0.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel file support

# Machine Learning - Core
scikit-learn>=1.3.0

# Machine Learning - Advanced (Optional but recommended)
xgboost>=2.0.0        # Gradient boosting
prophet>=1.1.5        # Time series forecasting

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.0

# Optional: Jupyter notebook support (if needed)
# jupyter>=1.0.0
# ipykernel>=6.25.0
